# docs/appApiV1_swagger.yaml
basePath: /api/app/v1
definitions:
  ErrorStruct:
    properties:
      error_code:
        type: integer
      error_message:
        type: string
    type: object
  domain.ClientGenderString:
    enum:
    - male
    - female
    type: string
    x-enum-varnames:
    - ClientGenderStringMale
    - ClientGenderStringFemale
  v1.clientLoginRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  v1.clientLoginResponse:
    properties:
      access_token:
        type: string
    type: object
  v1.clientRegisterRequest:
    properties:
      birthday:
        type: string
      email:
        type: string
      firstname:
        type: string
      gender:
        $ref: '#/definitions/domain.ClientGenderString'
      password:
        minLength: 6
        type: string
      surname:
        type: string
    required:
    - email
    - firstname
    - password
    - surname
    type: object
info:
  contact: {}
  description: Backend API for NNBlog Service
  title: Backend API
  version: "1.0"
paths:
  /clients/login:
    post:
      consumes:
      - application/json
      description: Авторизация
      parameters:
      - description: Авторизация
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.clientLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.clientLoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorStruct'
        "500":
          description: Internal Server Error
      summary: Авторизация
      tags:
      - Client
  /clients/ping:
    post:
      consumes:
      - application/json
      description: Ping
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "403":
          description: Forbidden
      security:
      - UserAuth: []
      summary: Ping
      tags:
      - Client
  /clients/register:
    post:
      consumes:
      - application/json
      description: Регистрация
      parameters:
      - description: Регистрация
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.clientRegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorStruct'
        "500":
          description: Internal Server Error
      summary: Регистрация
      tags:
      - Client
securityDefinitions:
  UserAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"